requirements.txt
altgraph==0.17.3
auto-py-to-exe==2.32.0
bottle==0.12.25
bottle-websocket==0.2.9
cffi==1.15.1
EasyProcess==1.1
Eel==0.14.0
entrypoint2==1.1
future==0.18.3
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.2
mss==7.0.1
numpy==1.24.2
opencv-python==4.7.0.72
pefile==2023.2.7
Pillow==9.4.0
pycparser==2.21
pygame==2.2.0
pyinstaller==5.8.0
pyinstaller-hooks-contrib==2023.0
pyparsing==3.0.9
pyscreenshot==3.0
pywin32-ctypes==0.2.0
whichcraft==0.6.1
zope.event==4.6
zope.interface==5.5.2
